==> src/01-EDA.Rmd <==
---
title: "EDA"
output:
  html_document:
    df_print: paged
  html_notebook: default
  pdf_document: default
---

## **Load Libraries**

```{r message=FALSE, warning=FALSE}
library(tidyverse)# data manipulation,create plots
library(DataExplorer) # correlation matrix
```

### **Load Data**




```{r message=FALSE, warning=FALSE}
df<-read.csv("houses_to_rent.csv")
```
```{r}
head(df)
```

```{r}
df<- df %>% select(-c(floor,animal,hoa,total,X,city,property.tax))
```


#### **Type of Data**

```{r}
df %>% select(fire.insurance,rent.amount) %>% summary()
```



We note that the price of fire insurance,floor and the price of rent. They are character type values, due to the fact that you have special characters, R standardizes and converts them to character type.




### **Convert variables to numeric**

```{r}
conv_to_numeric<-function(x){
  
 
  x<-gsub("[//R$,]","",x)   # eliminate special charterers
  x<-as.numeric(x) # transform to numeric data

  return(x) # return data
}

attach(df)
```


```{r}
fire_insurence<-sapply(fire.insurance,conv_to_numeric)
rent_amount<-sapply(rent.amount,conv_to_numeric)
```

```{r}
df<-df %>% 
  
  mutate(fire.insurance=fire_insurence,
         rent.amount=rent_amount)

# With the mutate function we make modifications to the data frame.
             
```





#### **Do furnished houses have a higher rental price compared to those that are not?**


```{r}
df %>% group_by(furniture) %>% summarise(rent_amount_mean=mean(rent.amount))
```


The average price of furnished houses is higher than those that are not.






```{r}
  
  histogram<-function(x,...){
    
    df %>%
      ggplot(aes(x=x,y=..density..)) +
      geom_histogram(color="black",fill="#FFF0C9") +
      geom_density(color="black",lwd=1) +
      geom_vline(aes(xintercept=mean(x),color="mean")) +
      geom_vline(aes(xintercept=median(x),color="median")) +
      labs(col="Stadistcs") +
      theme(legend.position = "top") +
      ...
    
  
  }
  
```
```{r}
rent_amount_histogram=histogram(df$rent.amount,labs(x="Rent Amount",title = "Rent Amount"))
area_hsitogram=histogram(df$area,labs(x="Area",title = "Area"))
fire_insurence_histogram=histogram(df$fire.insurance,labs(x="Fire Insurence",title = "Fire Insurence"))
rooms_histogram=histogram(df$rooms,labs(x="Rooms",title = "Rooms"))
bathroom_histogram=histogram(df$bathroom,labs(x="Bathroom",title = "Bathroom"))
parking_spces_histogram<-histogram(df$parking.spaces,labs(x="Parking Spaces",title="Parking Spaces"))

```

```{r}
scattter_plot<-function(x_feature,...){
 
  ggplot(data=df,aes(x=x_feature,y=rent.amount)) +
  geom_point(color="#77dd77",alpha=0.5) +
  theme_light() +
  geom_smooth(method = "lm",color="red") +
  ...
}
```

```{r}
area_scatter<-scattter_plot(df$area,labs(x="Area",y="Rent Amount"))
fire_scatter<-scattter_plot(df$fire.insurance,labs(x="Fire Insurence",y="Rent Amount"))
rooms_scatter<-scattter_plot(df$rooms,labs(x="Rooms",y="Rent Amount"))
bathroom_scatter<-scattter_plot(df$bathroom,labs(x="Bathroom",y="Rent Amount"))
parking_scatter<-scattter_plot(df$parking.spaces,labs(x="Parking Spaces",y="Charges"))
```

```{r message=FALSE, warning=FALSE}
library(gridExtra)
```




### **Histogram Plots**


```{r}
fire_insurence_histogram
```


For the variable fire insurence there is a high amount of abnormal values. There are very few cases where they exceed R $250.



```{r}
area_hsitogram
```


There is an important bias for the variable area. At first glance, it can be seen that most of the departments have an area less than and equal to square meter.



```{r}
rent_amount_histogram
```

It can be considered that most of the rental houses are around a price below R $15,000.



```{r}
rooms_histogram
```


Most rooms have a maximum amount of 7.




```{r}
bathroom_histogram
```


In most rental houses records have a maximum of 6 bathrooms.




```{r}
parking_spces_histogram
```

There are very few departments where it exceeds more than 6 places to park.



### **Scatter Plots**



```{r}
grid.arrange(area_scatter,
             fire_scatter,
             rooms_scatter,
             bathroom_scatter,parking_scatter)
             
```



The area variable apparently does not show any possible correlation with the rental price. Since it contains a good number of outliers, which causes the data to be skewed.



There is a clear correlation between the price of fire insurance with respect to the price of the rental house, since the more expensive the price of the apartment, the more you will have to pay to protect it against fire. These variables have a linear relationship, that is, they increase proportionally with another.





### **Correlation Matrix**





Shows the degree of relationship of the variables. They are measured from 0 to 1 if it is a positive correlation, otherwise it is measured from 1 to -1.


```{r}
plot_correlation(df,title = "Correelation Matrix")
```
### **Save Dataset**

```{r}
write.csv(df,"rent-amount-brazil.csv",row.names = FALSE)
```


### **Conclusion**



There is a strong presence of outliers in the data set. Especially for the area variable. For continuous variables, that is, those values with decimals, we can perform a logarithmic transformation to transform the outliers.


Variables such as the size of the apartment area, the number of bathrooms, number of bedrooms and fire insurance. It makes all the sense in the world for it to increase prices, since these qualities increase the size of houses. The higher the cost of the apartment, the higher the price of fire insurance, since you will have to cover more costs due to the proportion of the apartment.
