==> src/02-Feature_enginiering.Rmd <==
---
title: "Feature engineering"
output:
  html_document:
    df_print: paged
  html_notebook: default
  pdf_document: default
---

# **Load Libraries**

```{r message=FALSE, warning=FALSE}
library(tidyverse) # common libraries
```

### **Load Data**

```{r}
df<-read.csv("rent-amount-brazil.csv")
```

```{r}
head(df)
```



In EDA we conclude that the data set has too many outliers, which is why they require a type of processing.




I can think of 3 ways to deal with the situation.



* Remove outliers.
* Replace abnormal values by statistical measures.
* Perform a logarithmic transformation
* Calculate an upper range and based on the range replace outliers with normal values.




The first and second options are ruled out, since the data set has very few observations. And if we replace the outliers with statistical measures such as the mean or the median, we would be dirtying the data set, we would be affecting the distribution of the data.




We can combine the 3 and 4 option, we can establish an **upper range**. Based on the interval we create a criterion, we make a **sample** of random values that oscillate in those ranges. To be able to **replace** them and then perform a **logarithmic transformation** to smooth the data, improving their distribution.




```{r message=FALSE, warning=FALSE}
attach(df)
```

```{r}
upper_limit<-function(x,limit){
  mean<-mean(x) # calculate mean
  sd<-sd(x) # calculate standard division
  
  mean+limit*sd # calculate upper range 
}
```





```{r}

calculate_upper_limit<-function(feature){
  
  limits<-c(2,2.5,3,3.5,4)
  for(index in 1:5){
    
    print(paste("Upper Limit",limits[index]))
    print(upper_limit(feature,limits[index]))
    print("====================")
}
}

```



#### **Upper Limits**



##### **Rent Amount**

```{r}
calculate_upper_limit(feature = rent.amount)
```




The best upper limit is 3 which we can round up to R $15,000. Since there are more rental houses that are around those prices as we can see in the histogram.



#### **Fire Insurence**


```{r}
calculate_upper_limit(feature = fire.insurance)
```



With an upper limit of 4 it is good. 
With an upper limit of 4 it is good. We can round it up to R $250  since there are very few cases where the price of fire insurance exceeds that amount.



#### **Area**


```{r}
calculate_upper_limit(feature = area)
```



With an upper limit of 2.5 it is a good point. Since most of the departments do not exceed 1000 square meters.







```{r}
replace_outlires<-function(data,normal_sample,upper_limit){
  
  return(ifelse(data>upper_limit,normal_sample,data))
}
```

### **Normal Values Sample for Upper Limit**

```{r}
set.seed(2018) # we define random seed, so that the numbers generated by the sample do not vary.
sample_rent<-sample(14500:15000,size=25,replace = T)
sample_fire_insurence<-sample(240:250,size=10,replace = T)
sample_area<-sample(900:1000,size=25,replace = T)
```

```{r}
normal_values_rent<-replace_outlires(data=df$rent.amount,sample_rent,upper_limit = 15000)
normal_fire_insurence<-replace_outlires(data=df$fire.insurance,sample_fire_insurence,upper_limit = 250)
normal_area<-replace_outlires(data=area,sample_area,upper_limit = 1000)
```

## **Replace Outlires Values**

```{r}
df<- df %>%
  mutate(rent.amount=normal_values_rent) %>%
  mutate(fire.insurance=normal_fire_insurence) %>%
  mutate(area=normal_area)
```




With the mutate function we make modifications. We replace outliers with normal values, using a random sample.



##### **We check the changes**

```{r}
df %>% select(rent.amount,area,fire.insurance) %>% summary()
```

```{r}
df<- df %>% mutate(fire.insurance=ifelse(fire.insurance==3,5,fire.insurance))
```



We replaced the minimum price value of fire insurance by five real Brasilian.



```{r}
df %>% select(fire.insurance) %>% summary()
```





#### **logarithmic transformation**




```{r}
df<- df %>%
  
  mutate(area_log=log(area)) %>%
  mutate(fire.insurance_log=log(fire.insurance)) %>%
  mutate(rent.amount_log=log(rent.amount))

```




We perform logarithmic transformation to improve the distribution of continuous data.




```{r}
histogram<-function(x,...){
    
    df %>%
      ggplot(aes(x=x,y=..density..)) +
      geom_histogram(color="black",fill="#FFF0C9") +
      geom_density(color="black",lwd=1) +
      geom_vline(aes(xintercept=mean(x),color="mean")) +
      geom_vline(aes(xintercept=median(x),color="median")) +
      labs(col="Stadistcs") +
      theme(legend.position = "top") +
      ...
    
  }
```





```{r message=FALSE, warning=FALSE}
library(gridExtra)
```





```{r}
rent_amount_histogram=histogram(df$rent.amount,labs(x="Rent Amount",title = "Rent Amount"))
area_hsitogram=histogram(df$area,labs(x="Area",title = "Area"))
fire_insurence_histogram=histogram(df$fire.insurance,labs(x="Fire Insurence",title = "Fire Insurence"))

rent_amount_histogram_log=histogram(df$rent.amount_log,labs(x="Rent Amount",title = "Rent Amount Log"))
area_histogram_log=histogram(df$area_log,labs(x="Area",title = "Area Log "))
fire_insurence_histogram_log=histogram(df$fire.insurance_log,labs(x="Fire Insurence",title = "Fire Insurence Log"))

```




```{r}
grid.arrange(rent_amount_histogram,
             rent_amount_histogram_log,nrow=1)
         
```




```{r}
grid.arrange(fire_insurence_histogram,
             fire_insurence_histogram_log,nrow=1)
```


```{r}
grid.arrange(area_hsitogram,
             area_histogram_log,nrow=1)
```



We observed a great improvement in the distribution of the data.






```{r}
rooms_histogram<-histogram(df$rooms,labs(x="Rooms"))
bathroom_histogram<-histogram(df$bathroom,labs(x="Bathroom"))
parking_spaces_histogram<-histogram(df$parking.spaces,labs(x="Parking Spaces"))
```


```{r}
rooms_histogram
```

```{r}
bathroom_histogram
```


```{r}
parking_spaces_histogram
```

We can replace both variables, the values that are out of a normal range by the amount of 5, since it is a good confidence interval.


```{r}
replace_values<-function(x){
  
  ifelse(x>5,5,x)
  
}
```




```{r}
replace_rooms<-mapply(replace_values,rooms)
replace_bathrooms<-mapply(replace_values,bathroom)
replace_spaces_parking<-mapply(replace_values,parking.spaces)
```




```{r}
df<- df %>%
  
  mutate(replace_rooms=replace_rooms) %>%
  mutate(replace_bathrooms=replace_bathrooms) %>%
  mutate(replace_parking=replace_spaces_parking)
  
```




```{r}
rooms_hist_replace<-histogram(df$replace_rooms,labs(x="Rooms",title ="Replace Values"))
bathroom_hist_replace<-histogram(df$replace_bathrooms,labs(x="Bathroom",title = "Replace Values"))
parking_hist_replace<-histogram(df$replace_parking,labs(x="Parking Spaces",title = "Replace Values"))
```


```{r}
grid.arrange(rooms_histogram,rooms_hist_replace,nrow=1)
```
```{r}
grid.arrange(bathroom_histogram,bathroom_hist_replace,nrow=1)
```


```{r}
grid.arrange(parking_spaces_histogram,parking_hist_replace,nrow=1)
```



We finish with the cleaning of our data.


#### **We remove old variables**

```{r}
df<-df %>% 
  
  select(-rent.amount,-area,-fire.insurance) %>% 
  select(-rooms,-bathroom,-parking.spaces)
```


#### **Rename Features**


```{r}
df_clear<-df %>%
  
  rename(fire.insurance=fire.insurance_log,
         area=area_log,
         rent.amount=rent.amount_log) %>%
  
  rename(bathrooms=replace_bathrooms,
         parking.spaces=replace_parking,
         rooms=replace_rooms) 
```




```{r}
quality_departament<-cut(df_clear$rent.amount,breaks = 5,labels = c(1,2,3,4,5))
```





```{r}
df_clear<-df_clear %>% mutate(quality_departament=quality_departament)
```



```{r}
df_clear<- df_clear %>% mutate(quality_departament=as.numeric(quality_departament))
```


We can classify the degree of quality of the rental house. Generally the higher the price, the higher the quality of the house. We can group it from a range of 1 to 5, the higher the number, the greater the impact of the department.



```{r}
library(DataExplorer) # correlation matrix
```

```{r}
plot_correlation(df_clear,title = "Correlation Matrix")
```


## **Save Dataframe clear**

```{r}
write.csv(df_clear,"rent-amount-brazil-clear.csv",row.names = FALSE)
```
