==> src/04-Model_ideal.Rmd <==
---
title: "Model Ideal"
output:
  html_document:
    df_print: paged
  html_notebook: default
  pdf_document: default
---

## **Load Libraries**



```{r message=FALSE, warning=FALSE}
library(tidyverse) # common libraries
library(caret)
```



## **Load Data**



```{r message=FALSE, warning=FALSE}
df<-read.csv("rent-amount-brazil-clear.csv")
```








```{r}
df<-df %>% mutate(furniture=ifelse(furniture=="furnished",1,0))
```





### **Split Data**


```{r message=FALSE, warning=FALSE}
set.seed(2018) # random state

training.ids<-createDataPartition(df$rent.amount,p=0.7,list=F)

train_data<-df[training.ids,]
test_data<- df[-training.ids,]
```



## **Model Creation**


```{r}
lm<-lm(rent.amount~.,data=train_data)
```


### **MSE**



Measures the average error between the predicted value and the original.


#### **R²**



It measures the degree of fit of the model. The higher the value, the better the model fit.




```{r}
summary(lm)
```

The model has a good fit to the data. It can explain 96% of the cases, which can generalize very well.


```{r}
y_pred_test<-predict(lm,newdata = test_data)
```

```{r}
test_data<-test_data %>% mutate(y_pred=y_pred_test)
```







```{r}
test_data %>% summarise(rmse=RMSE(y_pred,rent.amount))
```





```{r}
test_data %>% summarise(mse=R2(y_pred,rent.amount))
```



It has similar metrics with the training data, indicating that it's not just good for the data it was trained on. If not, it is also suitable, for values that I have never seen.



### **Exponential transformation**



```{r}
test_data<-test_data %>%
  
  mutate(rent.amount=exp(rent.amount)) %>%
  
  mutate(predictions=exp(y_pred))
```



Let us remember that the rent.amount variable is in logarithmic scale and we perform the exponential transformation because it is its opposite operation.

### **Variable coefficients**



```{r}
coef(lm)
```


Coefficients determined by the Linear Regression model.



```{r}
test_data %>% 
  ggplot(aes(x=rent.amount,y=predictions))+
  geom_point(color="green",alpha=0.15) +
  geom_abline(intercept = 0,slope = 1,color="red",lwd=2) +
  theme_bw() +
  labs(title = "True Values vs Predicted Values",
       x="True values",y="Predicted Values")
```

```{r}
test_data %>% select(rent.amount,predictions) %>% head(20)
```


In most predictions, it gives predictions very close to the original value.



### **Save Model**

```{r}
saveRDS(lm,"lm_rent_amount.RDS")
```
